Deprecate list-snapper-snapshots and add get-snapper-root
The naming on list-snapper-snapshots is bad, as by the name you would
think you get a list of the snapper snapshots for that config, but
instead it returns the root directory of the .snapshots dir.

get-snapper-root is just a rename of that, keeping the old one around
incase someone is using different versions between sender and
receiver.