Only fetch the last snapshot in the backup function [16]
`p_snapshot` is only used in the backup function and there are edge case
where the source snapshots can be empty and you want to prune the
destination snapshots. Which will cause a bad array subscript. See
issue [0].

Fetching the last snapshot in the backup function guarantees that
there will be at least one source snapshot.

[0] https://github.com/plattfot/baksnapper/issues/16