bugfix: Use a binary map to track common snapshots [#27]
When sending all snapshots and there are holes in the source
snapshots, it will result in it using an older parent for sending
incremental snapshots.  Causing snapshots to be bigger than necessary.

This commit fixes that by changing how it represent the common
snapshots.  Instead of having them in an array.  It now uses a sparse
array where the index to the array is the snapshot number and the
value is a small bitmap that represent the existence of the snapshot
at the two locations.

This becomes useful when sending all snapshots as it can now merge the
common and source only snapshots into one array and from there get the
right parent.

Resolves #27