Migrate ChangeLog to a Keep A Changelog style [#40]
A lot easier to work with and easier to get a quick overview of the
changes.

This one does not 100% follow the Keep A Changelog recommendations as
the unreleased changes are kept in separate files in CHANGELOG.d.
That allows for a much nicer development experience as Pull Requests
will not conflict with each other.

Resolves #40.